backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "kafka-python (>=2.0.4,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "confluent-kafka (>=2.9.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false